requirements.txt
aniso8601==9.0.1
attrs==21.4.0
Bottleneck @ file:///C:/ci/bottleneck_1648010904582/work
certifi==2021.10.8
charset-normalizer==2.0.12
click @ file:///C:/ci/click_1646038595831/work
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
easyocr==1.4.2
Flask @ file:///opt/conda/conda-bld/flask_1648041541647/work
flask-restx==0.5.1
idna==3.3
imageio==2.19.0
itsdangerous @ file:///tmp/build/80754af9/itsdangerous_1621432558163/work
Jinja2 @ file:///opt/conda/conda-bld/jinja2_1647436528585/work
jsonschema==4.5.1
MarkupSafe @ file:///C:/ci/markupsafe_1621528502553/work
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184308/work
mkl-service==2.4.0
networkx==2.8
numexpr @ file:///C:/ci/numexpr_1640704337920/work
numpy==1.22.3
opencv-python-headless==4.5.4.60
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas @ file:///C:/ci/pandas_1650373458095/work
Pillow==9.1.0
pyparsing==3.0.8
pyrsistent==0.18.1
python-bidi==0.4.2
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
requests==2.27.1
scikit-image==0.19.2
scipy==1.8.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
tifffile==2022.5.4
torch==1.11.0
torchvision==0.12.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug @ file:///opt/conda/conda-bld/werkzeug_1645628268370/work
wincertstore==0.2
